requirements.txt
Flask==1.0.2
requests==2.20.1
humanize==0.5.1
python-dateutil==2.7.5
pytz==2018.7
Babel==2.6.0
